web/requirements.txt
flask
markupsafe
joblib
shap
numpy
pandas
python-dotenv
google-genai